pyproject.toml
[project]
name = "incident-triage-env"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai"
]
